+++ quill-core-portable/src/main/scala/io/getquill/norm/ApplyMap.scala
@@ -55,6 +55,7 @@ object ApplyMap {
     def unapply(ast: Ast): Option[(Ast, Ident, Ast)] =
       ast match {
         case Map(a: GroupBy, b, c)              => None
+        case Map(a: FlatJoin, b, c)             => None // FlatJoin should always be surrounded by a Map
         case Map(a, b, InfixedTailOperation(c)) => None
         case Map(a, b, c)                       => Some((a, b, c))
         case _                                  => None
@@ -66,6 +67,7 @@ object ApplyMap {
 
       case Map(a: GroupBy, b, c) if (b == c)    => None
       case Map(a: Nested, b, c) if (b == c)     => None
+      case Map(a: FlatJoin, b, c) if (b == c)   => None // FlatJoin should always be surrounded by a Map
       case Nested(DetachableMap(a: Join, b, c)) => None
 
       //  map(i => (i.i, i.l)).distinct.map(x => (x._1, x._2)) =>